Revert "fix: remove bigint from varint"
This reverts commit 3e5e372ff2d67f79f744fbd8b0c9c64b138c963a.

+++ app/src/main/java/protobuf/magic/protobuf/BufferReader.java
@@ -1,5 +1,6 @@
 package protobuf.magic.protobuf;
 
+import java.math.BigInteger;
 import javax.naming.InsufficientResourcesException;
 import protobuf.magic.struct.VarintResult;
 
@@ -18,7 +19,7 @@ public class BufferReader {
     this.offset = 0;
   }
 
-  public Long readVarInt() {
+  public BigInteger readVarInt() {
     VarintResult result = VarintUtils.decodeVarint(buffer, offset);
     offset += result.length();
     return result.value();
